requirements.txt
altgraph==0.17.2
certifi==2021.5.30
charset-normalizer==2.0.6
configparser==5.0.2
future==0.18.2
idna==3.2
MouseInfo==0.1.3
pefile==2021.9.3
Pillow==8.3.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyMsgBox==1.0.9
pynput==1.6.8
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
PySide6==6.1.3
pytesseract==0.3.8
pytweening==1.0.4
pywin32==301
pywin32-ctypes==0.2.0
requests==2.26.0
shiboken6==6.1.3
six==1.16.0
urllib3==1.25.8
